requirements.txt
requests
beautifulsoup4
anthropic
pyyaml
python-dotenv
